rethnet_eth::remote: mv methods::eth to methods
